Move `APS_REPLY_TIMEOUT*` to const
